fix implicit cast warnings
(Implicit conversion turns floating-point number into integer: 'double' to 'uint64_t' (aka 'unsigned long long'))

+++ Lumberjack/DDFileLogger.m
@@ -681,7 +681,7 @@
 	});
 	#endif
 	
-	uint64_t delay = [logFileRollingDate timeIntervalSinceNow] * NSEC_PER_SEC;
+	uint64_t delay = (uint64_t)([logFileRollingDate timeIntervalSinceNow] * NSEC_PER_SEC);
 	dispatch_time_t fireTime = dispatch_time(DISPATCH_TIME_NOW, delay);
 	
 	dispatch_source_set_timer(rollingTimer, fireTime, DISPATCH_TIME_FOREVER, 1.0);
